# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yarn Dyeing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      tags: [orders]
      operationId: listOrders
      summary: List yarn orders with optional filters
      parameters:
        - name: factoryId
          in: query
          required: false
          schema:
            type: integer
        - name: buyerName
          in: query
          required: false
          schema:
            type: string
        - name: color
          in: query
          required: false
          schema:
            type: string
        - name: orderType
          in: query
          required: false
          schema:
            type: string
            enum: [Sample, Bulk]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Received, Lab Dip, Dyeing Running, Finishing, Ready, Delivered]
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags: [orders]
      operationId: createOrder
      summary: Create a new yarn order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags: [orders]
      operationId: getOrder
      summary: Get a yarn order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [orders]
      operationId: updateOrder
      summary: Update a yarn order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    delete:
      tags: [orders]
      operationId: deleteOrder
      summary: Delete a yarn order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /orders/{id}/status:
    post:
      tags: [statusHistory]
      operationId: updateOrderStatus
      summary: Update order production status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdate'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # ── PHOTOS ─────────────────────────────────────────────────────────────────
  /orders/{id}/photos:
    get:
      tags: [photos]
      operationId: listOrderPhotos
      summary: List photos for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Photos list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderPhoto'
    post:
      tags: [photos]
      operationId: addOrderPhoto
      summary: Add a photo record to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoInput'
      responses:
        '201':
          description: Photo added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPhoto'
  /orders/{id}/photos/{photoId}:
    delete:
      tags: [photos]
      operationId: deleteOrderPhoto
      summary: Delete a photo from an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'

  # ── DASHBOARD ──────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth schemas
...
    Order:
      type: object
      required: [id, orderNo, buyerName, orderType, yarnType, color, quantityKg, status, receiveDate, createdAt]
      properties:
        id:
          type: integer
        orderNo:
          type: string
        buyerName:
          type: string
        factoryId:
          type: [integer, 'null']
        factoryName:
          type: [string, 'null']
        orderType:
          type: string
          enum: [Sample, Bulk]
        yarnType:
          type: string
        color:
          type: string
        quantityKg:
          type: number
        receiveDate:
          type: string
          format: date
        deliveryDate:
          type: [string, 'null']
          format: date
        status:
          type: string
          enum: [Received, Lab Dip, Dyeing Running, Finishing, Ready, Delivered]
        remarks:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/OrderPhoto'
            statusHistory:
              type: array
              items:
                $ref: '#/components/schemas/StatusHistoryItem'
    OrderListResponse:
      type: object
      required: [orders, total, page, pageSize]
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    OrderInput:
      type: object
      required: [buyerName, orderType, yarnType, color, quantityKg, receiveDate]
      properties:
        orderNo:
          type: string
        buyerName:
          type: string
        factoryId:
          type: [integer, 'null']
        orderType:
          type: string
          enum: [Sample, Bulk]
        yarnType:
          type: string
        color:
          type: string
        quantityKg:
          type: number
        receiveDate:
          type: string
          format: date
        deliveryDate:
          type: string
          format: date
        remarks:
          type: string
    OrderUpdate:
      type: object
      properties:
        buyerName:
          type: string
        factoryId:
          type: [integer, 'null']
        orderType:
          type: string
          enum: [Sample, Bulk]
        yarnType:
          type: string
        color:
          type: string
        quantityKg:
          type: number
        receiveDate:
          type: string
          format: date
        deliveryDate:
          type: string
          format: date
        status:
          type: string
          enum: [Received, Lab Dip, Dyeing Running, Finishing, Ready, Delivered]
        remarks:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Received, Lab Dip, Dyeing Running, Finishing, Ready, Delivered]
        remarks:
          type: string
    StatusHistoryItem:
      type: object
      required: [id, orderId, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        status:
          type: string
        remarks:
          type: [string, 'null']
        updatedBy:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time

    # Photo schemas
    OrderPhoto:
      type: object
      required: [id, orderId, photoUrl, photoType, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        photoUrl:
          type: string
        photoType:
          type: string
          enum: [Challan, Yarn Photo, Shade Card, Delivery Evidence]
        createdAt:
          type: string
          format: date-time
    PhotoInput:
      type: object
      required: [photoUrl, photoType]
      properties:
        photoUrl:
          type: string
        photoType:
          type: string
          enum: [Challan, Yarn Photo, Shade Card, Delivery Evidence]

    # Dashboard schemas
...