# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderOrderType } from './orderOrderType';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNo: string;
  buyerName: string;
  /** @nullable */
  factoryId?: number | null;
  /** @nullable */
  factoryName?: string | null;
  orderType: OrderOrderType;
  yarnType: string;
  color: string;
  quantityKg: number;
  receiveDate: Date;
  /** @nullable */
  deliveryDate?: Date | null;
  status: OrderStatus;
  /** @nullable */
  remarks?: string | null;
  createdAt: Date;
}