# lib/api-zod/src/generated/types/statusHistoryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusHistoryItem {
  id: number;
  orderId: number;
  status: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  updatedBy?: string | null;
  createdAt: Date;
}